requirements.txt
streamlit
pdfplumber
langchain
openai
python-dotenv
